fix up about us page with grid.
